requirements.txt
# LangChain 核心框架
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23

# 工具依赖
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0

# 数据模型
pydantic>=2.0.0
